// src/modules/ticket-soporte-historial/domain/enums/ticket-soporte-historial-tipo.enum.ts
export enum TicketSoporteHistorialTipo {
  CREADO = 'CREADO',
  ACTUALIZADO = 'ACTUALIZADO',
  ESTADO_CAMBIADO = 'ESTADO_CAMBIADO',
  PRIORIDAD_CAMBIADA = 'PRIORIDAD_CAMBIADA',
  ASIGNACION_CAMBIADA = 'ASIGNACION_CAMBIADA',
  CANCELADO = 'CANCELADO',
  REABIERTO = 'REABIERTO',
  FIJADO = 'FIJADO',
  DESFIJADO = 'DESFIJADO',
}